experimental feature: a harsh mass-redirection script that can be used to temporarly turn off (by redirecting them) a user's subdomains (all of them)
temporarly, because a SQL dump is showed to allow the admin to restore the
previous state. 

experimental, because the SQL dump should probably be hidden from the admin and
that this whole thing should probably be called something else than
"deactivate" and might be better somewhere else than in adm_list.php.

harsh, because there could be other ways to deal with evil users. this is the
best solution I have found for accounts that are victims of spambots.