stop overwriting named.conf
starting this release, only named.conf.options will be replaced (as we
need to modify the options {} block and can't just add another one). We
still install a named.conf, a fresh version from Lenny this time, so
that we get a clean start.

Also get rid of the bind_internal parameter that is not really used or
recommended, as it allows for authoritative nameservers to also serve
recursive queries (which is bad practice).

See: #1025
Closes: #1104