regroup TLS options and modernize them
Options that are related to TLS are currently split into two groups in
the template, and this makes it hard to understand what is happening.

Some changes to configuration are necessary since a lot of time has
passed since the latest changes to this template. Some configuration is
actually plain wrong, and some other is doing what the documentation
recommends against. Following are some details about why options were
changed:

 * smtpd_use_tls and smtp_use_tls are deprecated. We should use instead
smtp(d)_tls_security_level.
   * smtpd_tls_security_level should be set to "may" in order to use
encryption opportunistically when delivering emails to other MTAs
whenever they advertise that they can do it.

 * smtpd_tls_dcert_file and smtpd_tls_dkey_file exist specifically for
certificates that use DSA key pairs. This is not the norm anymore and
RSA has been the default for many years now. We should use the more
generic option smtpd_tls_cert_file.
   * Since alternc generates a single file apache.pem that contains the
certificate and the private key, we don't actually need to specify
smtpd_tls_key_file.

 * smtpd_tls_CApath restricts postfix to using *only* the
system-provided CAs. This might not be what some folks want to do and
the restriction that the template imposes is useless.

 * we should *not* set smtp_tls_{dcert,dkey,cert,key}_file ! This is
setting up postfix to use a *client* certificate when contacting other
servers for delivery to other MTAs. This is very probably not what most
ppl want to be doing, and if some users actually want to do this because
for example the MTA configured by alternc needs to authenticate to a
remailer to get mail out to the world, then they will *not* want to use
the same certificate/private key pair than the one that's used on the
mail submission side (smtpd_*).

 * smtpd_tls_auth_only, when set to "no" permits clients to send their
credentials unencrypted! This is not a good default if we expect to be
using TLS

 * smtp(d)_tls_protocols should exclude deprecated cipher "families"
instead of including some of them, as is recommended in
http://www.postfix.org/postconf.5.html#smtpd_tls_protocols

 * smtpd_sasl_auth_enable is there twice