user cron jobs that fall outside the max concurrent limit cause an error
The $urls variable is an array of arrays. There was a sub-element access
that was left out in the part of the rolling_curl function that starts
user cron jobs that are waiting in queue because there were too many
cron jobs to process at once.

Arguably the code that starts a curl request should be split out into a
function and reused, but that change is more complex than a one-line fix
to an actual bug.