Order enumerated mails for a domain by ascending alphabetic order
Fixes #341

When applying LIMIT X, Y the ordering before the limit and offset do
not seem to be guaranteed. For example, if you have a large number of
e-mail addresses, and you page between the same e-mail address can appear twice.

The case where this was happening there were 2-3 mail boxes and ~90 aliases.

I'm not sure if this tied somehow to the database version used, but making the
desired ordering explicit ensures that all mails will eventually be shown and
shouldn't be shown twice.