Fix #413: Add return values for bind update domains hooks
hook_updatedomains_dns_{add,del} have their return values checked in
m_dom::update_domains() to set the dns_result column.

When no return value is specified, the value is actually NULL and doesn't
get mapped to an integer value that is expected by the dns_result column.

As a result, the query fails silently and the DNS_ACTION column is never
updated. This causes update_domains to retry the same modification on
every run.