Limit incrond scripts to watching a specific file
When they both watch the same directory, when a file is touched - eg,
/run/alternc/incron/inotify_do_action.lock both scripts are started
running in children of the top-level /usr/sbin/incrond script. When
those scripts finish, one or the other tends to start watching for all
the incron tables again (created more and more children each time).

Maybe it's a bug in incron.

Test:
 * Check number of running processes with ps faux | grep incron
 * Do an action, eg. chmod a folder in the web interace
 * Re-check the number of running processes, they should stay the same
 not increase