empty passwords, when permitted, should not be checked against any policy
Currently, when you create or edit an email address that is "not local"
(e.g. an alias) you get some beautiful error messages:

    Warning: strpos(): Empty needle in
/usr/share/alternc/panel/class/m_admin.php on line 1525
    Warning: strpos(): Empty needle in
/usr/share/alternc/panel/class/m_admin.php on line 1525
    Warning: strpos(): Empty needle in
/usr/share/alternc/panel/class/m_admin.php on line 1525

This is because one of the policy checks forgot to check if we are in
the case of an empty password when it is permitted to be empty.

but ... let's simplify this function a bit. instead of checking for this
case on all policy checks, we should short-circuit the function before
any policy check is done and just return that all is fine if the
password is empty and it is a case where it's permitted. the code is
easier to read this way and we won't forget to check for this if we add
more policy checks on the password.