user piwik account quota instead of hardcoded value
The code is already getting the value of the piwik quota ... so why is
it not using said quota?

There was also another bug in the code: if a quota issue was detected,
the piwik account creation was still proceeding nonetheless. This change
makes it so that an account is not created if a quota issue is detected.